\name{KMEANS599}
\alias{KMEANS599}
\title{K-Means}
\description{Calls code for \code{\line{KMEANS599} function that computs kmeans clustering for an input feature matrix.}
\usage{KMENS(data.matrix, 
    K = number.of.clusters,
    max.iterations = max.iterations)}

  \item{data.matrix}{Matirx of numeric data.}
  \item{K}{Maximum number of principal components to compute}


\value{List output from \code{\link{KMEANS599} which represents a kmeans clustering model.}

\author{Jared Benajmin Silverstein}
